pyproject.toml
[project]
name = "exp"
version = "0.1.0"
description = "Unified Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.118.0",
    "fastmcp>=2.12.4",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "mcp>=1.16.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "sse-starlette>=3.0.2",
    "uvicorn>=0.37.0",
    "websockets>=15.0.1",
]
